renamed ItemsService test to match src file name
